Refactor: move Metal preludes to separate file, extract Metal-only tests
- Move all 5 Metal prelude strings to slang-emit-metal-prelude.cpp
  (following slang-emit-hlsl-prelude.cpp pattern)
- Rename cuda_spirv_metal to cuda_metal_spirv (alphabetic order)
- Extract Metal fill test to metal-fill.slang (no shared code with VK/CUDA)
- Extract Metal groupshared test to metal-load-store-groupshared.slang
- Revert fill.slang and load-store-groupshared.slang to original (VK/CUDA only)
- Drop redundant CHECK_METAL in array.slang and struct.slang (reuse CHECK)

Made-with: Cursor
