+++ docs/command-line-slangc-reference.md
@@ -1401,7 +1401,7 @@ A capability describes an optional feature that a target may or may not support.
 * `cpp_cuda_spirv_llvm` 
 * `cpp_cuda_metal_spirv` 
 * `cuda_spirv` 
-* `cuda_spirv_metal` 
+* `cuda_metal_spirv` 
 * `cpp_cuda_glsl_spirv` 
 * `cpp_cuda_glsl_hlsl` 
 * `cpp_cuda_glsl_hlsl_llvm`
