+++ docs/user-guide/a3-02-reference-capability-atoms.md
@@ -1171,11 +1171,11 @@ Compound Capabilities
 `cuda_hlsl_spirv`
 > CUDA, HLSL, SPIRV code-gen targets
 
+`cuda_metal_spirv`
+> CUDA, Metal and SPIRV code-gen targets
+
 `cuda_spirv`
 > CUDA and SPIRV code-gen targets
-
-`cuda_spirv_metal`
-> CUDA, SPIRV and Metal code-gen targets
 
 `descriptor_handle`
 > Targets that support DescriptorHandle types for bindless descriptor access.
